CLIENT-GAME: Add rect animation MINOR
