+++ src/Client/Systems/ClientSystems.cpp
@@ -6,6 +6,9 @@
 */
 
 #include "ClientSystems.hpp"
+#include "EventsSystems.hpp"
+#include "GraphicSystems.hpp"
+#include "Systems.hpp"
 
 namespace Systems {
     std::array<std::vector<std::function<void(std::size_t, std::size_t)>>, 3>
