+++ src/Client/Systems/ClientSystems.hpp
@@ -8,9 +8,9 @@
 #pragma once
 
 #include <array>
-#include "EventsSystems.hpp"
-#include "GraphicSystems.hpp"
-#include "Systems.hpp"
+#include <cstddef>
+#include <functional>
+#include <vector>
 
 namespace Systems {
     std::array<std::vector<std::function<void(std::size_t, std::size_t)>>, 3>
