+++ src/Client/Systems/Graphic/AudioSystems.hpp
@@ -2,7 +2,7 @@
 ** EPITECH PROJECT, 2023
 ** R-Bus
 ** File description:
-** GraphicSystems
+** AudioSystems
 */
 
 #pragma once
@@ -13,7 +13,10 @@
 
 namespace Systems {
     namespace GraphicSystems {
+        void soundEffectPlayer(std::size_t /*unused*/, std::size_t /*unused*/);
+        void musicPlayer(std::size_t /*unused*/, std::size_t /*unused*/);
+        void playSoundWithKey(std::size_t, std::size_t);
         std::vector<std::function<void(std::size_t, std::size_t)>>
-        getGraphicsSystems();
+        getAudioSystems();
     } // namespace GraphicSystems
 } // namespace Systems
