+++ src/Client/Systems/Graphic/CMakeLists.txt
@@ -10,4 +10,7 @@ target_sources(
     ${PROJECT_NAME_CLIENT}
     PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/GraphicSystems.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/AudioSystems.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/SpriteSystems.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/TextSystems.cpp
 )
