+++ src/Client/Systems/Graphic/TextSystems.hpp
@@ -2,7 +2,7 @@
 ** EPITECH PROJECT, 2023
 ** R-Bus
 ** File description:
-** GraphicSystems
+** TextSystems
 */
 
 #pragma once
@@ -13,7 +13,8 @@
 
 namespace Systems {
     namespace GraphicSystems {
+        void textRenderer(std::size_t /*unused*/, std::size_t /*unused*/);
         std::vector<std::function<void(std::size_t, std::size_t)>>
-        getGraphicsSystems();
+        getTextSystems();
     } // namespace GraphicSystems
 } // namespace Systems
